FOGL-737 Added purge by size

+++ python/foglamp/tasks/purge/purge.py
@@ -48,6 +48,12 @@ class Purge(FoglampProcess):
             "type": "integer",
             "default": "72"
         },
+        "size": {
+            "description": "Maximium size of data to be retained, the oldest data will be removed to keep below this size," +
+                           "unless retained. (in Hours)",
+            "type": "integer",
+            "default": "1000000"
+        },
         "retainUnsent": {
             "description": "Retain data that has not been sent to any historian yet.",
             "type": "boolean",
@@ -71,12 +77,10 @@ class Purge(FoglampProcess):
         """" INSERT into log table values """
         payload = PayloadBuilder().INSERT(code='PURGE', level=level, log=log).payload()
         self._storage.insert_into_tbl("log", payload)
-        if level == 0:
-            self._logger.info("PURGED SUCCESSFULLY: ", log)
-        elif level == 2:
-            self._logger.warning("PURGED REMOVED UNSENT ROWS: ", log)
+        if level == 1:
+            self._logger.error("Purge failed: ", log)
         else:
-            self._logger.error("ROWS FAILED TO REMOVE: ", log)
+            self._logger.info("Purge: ", log)
 
     def set_configuration(self):
         """" set the default configuration for purge
@@ -111,22 +115,39 @@ class Purge(FoglampProcess):
 
 
         flag = "purge" if config['retainUnsent']['value'] == "False" else "retain"
-        result = self._readings_storage.purge(age=config['age']['value'], sent_id=last_id, flag=flag)
+        if config['age']['value'] != 0:
+            result = self._readings_storage.purge(age=config['age']['value'], sent_id=last_id, flag=flag)
 
-        """ Default to a warning
-            TODO Make a common audit trail class for inserting itno this log table
-        """
-        error_level = 4
+            """ Default to a warning
+                TODO Make a common audit trail class for inserting itno this log table
+            """
+            error_level = 4
 
-        if "message" in result.keys() and "409 Conflict" in result["message"]:
-            """ Message has become a failure """
-            error_level = 1 
-        else:
-            total_count = result['readings']
-            total_rows_removed = result['removed']
-            unsent_rows_removed = result['unsentPurged']
-            unsent_retained = result['unsentRetained']
+            if "message" in result.keys() and "409 Conflict" in result["message"]:
+                """ Message has become a failure """
+                error_level = 1 
+            else:
+                total_count = result['readings']
+                total_rows_removed = result['removed']
+                unsent_rows_removed = result['unsentPurged']
+                unsent_retained = result['unsentRetained']
 
+        if config['size']['value'] != 0:
+            result = self._readings_storage.purge(size=config['size']['value'], sent_id=last_id, flag=flag)
+
+            """ Default to a warning
+                TODO Make a common audit trail class for inserting itno this log table
+            """
+            error_level = 4
+
+            if "message" in result.keys() and "409 Conflict" in result["message"]:
+                """ Message has become a failure """
+                error_level = 1 
+            else:
+                total_count += result['readings']
+                total_rows_removed += result['removed']
+                unsent_rows_removed += result['unsentPurged']
+                unsent_retained += result['unsentRetained']
 
         end_time = time.strftime('%Y-%m-%d %H:%M:%S.%s', time.localtime(time.time()))
